Rewrite much of MHD TP with a new execution model
There is a comment in the mhd3d_tbp source file
describing the new rational. Basically, we're
restoring aspects of the functional execution
model, but replacing the concept of simulation
state with state definitions, that are submitted
to the runtime as they come through.

The HDF5 and VTK output functions are deleted right
now, they'll be recovered from the git history soon.