Reverse the way edges are removed as computables are evaluated
The old approach removed the outgoing node edges as soon as
the node is evaluated. The new approach waits until each
downstream node is evaluated to remove the edge. This is to
accommodate MPI execution strategy, where by mapping over an
executable that has a value, we still need to know its
downstream nodes so the resulting value can be sent to the
delegated MPI rank.