Protect against tag overflows in MPI wrappers
This is done in anticipation of using MPI tags to identify
serialized tasks. This solution may not scale in the future,
since some systems may not support tags larger than ~32,000.
But it is a convenient solution for now. The MPI wrapper
classes now check that the integer tag does not exceed the
maximum tag, MPI_TAG_UB, supported by the MPI implementation,
to make sure that if this ever becomes a limitation, the code
will fail with an appropriate error.