Significant refactor of computable and minidisk problem
This commit reworks the minidisk scheme to simplify the
execution graph. The performance on a single node is
the same as before, but the graph delegation is now
symmetric and should hopefully yield much better MPI
scaling. The MPI execution strategy is temporarily
disabled, to simplify refactoring of the computable
support functions.